requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Performance optimization - semantic search
sentence-transformers==2.7.0
faiss-cpu==1.7.4
numpy==1.24.3

